arm64: dts: imx8mq-librem5: Enable lm3560 LED
This enables torch and flash. We use the chips
default values as max for the moment. This can
be tweaked furthers as needed.

A single pinctrl group is enough and we can drop the regulator
since it's in fact a GPIO tied directly to the lm3560.

Maximum led current is derived from experiments on dogwood.

Signed-off-by: Guido Günther <agx@sigxcpu.org>