hi846: remove digital gain v4l control
this makes sense because digital gain is now used for white balance
calibration. Overwriting it would undo that calibration.

there are others drivers too that expose analogue gain only, as does
the android driver that I loosely based this driver onto.

We never used digital gain in any application so far on the Librem 5
so this is no visible change.