hack: arm64: dts: imx8mq-librem5: add 4G_WAKE as gpio-key
Ideally this (and the ones we describe as leds) line should be handled
by a driver I think, but this is a working workaround we do for a long
time - on the devkit as well.

Describe the 4G_WAKE line (on the modem side mostly called WOWWAN) that
gets pulled down (for about 1 second) on incoming call or SMS.

At least it should. We see modem devices that don't set this line.
So I think it makes sense to add this description now. It eases
testing of modems. For example, do:

	sudo evemu-record

and select "gpio-keys" and call the phone from another phone. When
it's ringing, it should now print the "KEY_PHONE" key got "pressed".