builddeb: Uncompress the kernel
This allows our uboot to boot the kernel directly and we
avoid having flash-kernel uncompressing it. While this
might look like a waste of space it isn't since this
makes us save

- vmlinuz (uncompressed)
- vmlinuz.bak (uncompressed)
- initrd.img
- initrd.img.bak

in /boot. It also avoids the uncompress itself which is more susceptible
to inconsistent state.

This makes us more align with Debian and once uboot allows to boot
compressed kernels we can just drop this patch without further changes
needed in e.g. flash-kernel.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>