power: supply: max17042_battery: Use Current register in get_status
It can take a while for AvgCurrent to adjust after (un)plugging
the charger. Use the instantaneous value in order to not confuse the
userspace.

While at that, don't do unit conversion of the read value. The current
code was prone to overflows and we only care about the sign anyway.

Signed-off-by: Sebastian Krzyszkowiak <sebastian.krzyszkowiak@puri.sm>