hack: usb: add quirk to skip reset_resume in hub_activate()
For devices with AVOID_RESET_RESUME set, dont set reset_resume in
hub_activate() but (from hub_resume()) but let it be set later via
port_resume() only.

This is a hacky workaround to improve stability when a permanently
connected device acts like it's disconnected from the hubs' port.

Since we see this for the BM818 modem, enable the quirk there.